requirements_testing.txt
elasticsearch
findspark
flask
fluent-logger
neo4j
neobolt
nox
pandas
py2neo
pyspark
pytest
redis
requests
waitress